fix: disable course contact buttons after 30 days (#401)
Closes: https://github.com/corona-school/project-user/issues/994

## ℹ️ Info
No timestamp is set when a subcourse is canceled. Therefore, the system
checks here whether the course has been canceled; if not, it looks at
the last course date to see whether it was 30 days ago. If one of the
cases applies, the buttons for the chat and the video meeting on the
course details page are hidden.

## ✅ Done 
- [x] check if subcourse is active
- [x] fix next appointment for students to join videomeeting 4h earlier