Feat: Add settings on the screening blocker (#508)
## Ticket

https://github.com/corona-school/project-user/issues/1159

## What was done?

- Added a new `RequireScreeningSettingsDropwon` that will appear on the
screening blocker. This component allows the user to deactivate their
account, or change their notifications preferences _(via a new
`NotificationsPreferencesModal` which just re-uses the existing
component but in a modal)_.
- Added links to Datenschutz and Impressum, similar to how we do it in
the login page
- Slightly updated how a "full" modal look like, so it at least have
some space on the sides which looks a bit better. I checked and this
modal size was only being used on the screeners section, and it's still
looking good with the new changes.

Here is how it looks like:


https://github.com/corona-school/user-app/assets/161815068/22bd1564-7727-433b-810d-24ba4939fd05