feat: Cache GraphQL queries (#533)
Basic idea: Keep a cache for each query's result, if a query is fetched
again, directly return the value from cache but also refetch the data in
the background, then update the cache and the page with that. This way,
the app more or less stays consistent and stays "as is" on devices with
low-latency. On a bad connection (or even offline) one can still
navigate around the app and the old data is displayed.

<img width="824" alt="Screenshot 2024-05-09 at 00 27 06"
src="https://github.com/corona-school/user-app/assets/23138527/e9cd8eac-fe24-4449-8cf0-b09c4e8c38ae">

---------

Co-authored-by: Jonas <jonas@wilms.ninja>