fix: Skip cache for appointment queries
Unfortunately pagination in GraphQL is terribly broken, and we update the query result of an existing query with fetchMore(...) passing in other variables, without changing the original variables of the query. Thus this messes up the cache.

Unfortunately for some reason cachePolicy: "no-cache" does not work and the cache is still taken :/ Thus adding a hack to skip the cache, by adding NO_CACHE in the query name