fix: Adjusted script to not send specific symbols that may prevent translations (#770)
## What was done?

https://lern-fair.slack.com/archives/C01548FCGBY/p1738678064969759
It seems that the combination of `:` + `{` prevents Weglot from
translating some texts. They mention on the documentation that there are
some special characters that cause this behavior _(I couldn't find the
list 🤷‍♂️)_

So I did the following:

- Replace `{{variable}}` with `[[index]]` before sending the text to
Weglot
- Replace again `[[index]]` with `{{variable}}` on the response so we
don't have to do any extra changes in the code and it works fine with
i18next
- Also adjusted a bit the regex to get the templates to include
uppercases. There are some translation keys that have special characters
like `{{subjectSchüler}}` that are not considered by the regex, we
should adjust that in a separate PR.