fix: Remove useless Backend Requests
Some time ago, the RecommendationsContext was introduced, adding a NotificationPreferences query to every page load - even if the user did not log in yet. Its a bit ugly that we load this globally now, but now at least we wait for the user to be authenticated first, and avoid the request otherwise.

Also avoid determining the session user on first page load - we just shuffled a new session token, so it cannot be associated with a session anyways.

This should reduce the amount of needless AuthenticationError we produce in the Backend.